requirements.txt
anyio==4.13.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
dnspython==2.8.0
elastic-transport==9.2.1
elasticsearch==9.3.0
Flask==3.1.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.7
